requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.28
botocore==1.29.28
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.3
djangorestframework==3.14.0
gevent==22.10.2
greenlet==2.0.1
jmespath==1.0.1
kombu==5.2.4
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.2
